# doc/openapi.json
{"openapi":"3.0.2","info":{"title":"ArtCafe API","description":"REST service for ArtCafe","contact":{"name":"ArtCafe","url":"http://x-force.example.com/contact/","email":"dp@x-force.example.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"0.0.1"},"paths":{"/api/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Login User and return JWT token","operationId":"login_api_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register User and return JWT token","operationId":"register_api_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/logout":{"post":{"tags":["Authentication"],"summary":"Logout","description":"Logout User, revoke JWT token","operationId":"logout_api_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/user/{user_id}":{"get":{"tags":["User"],"summary":"Get User","description":"Get details of a user by id. Requires authentication. Only if role is Admin or user_id is the same as the user_id of the authenticated user.","operationId":"get_user_api_user__user_id__get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfoModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"LoginModel":{"title":"LoginModel","required":["email","password"],"type":"object","properties":{"email":{"title":"Email","type":"string"},"password":{"title":"Password","type":"string"}}},"RegisterModel":{"title":"RegisterModel","required":["email","password","name","role"],"type":"object","properties":{"email":{"title":"Email","type":"string"},"password":{"title":"Password","type":"string"},"name":{"title":"Name","type":"string"},"role":{"title":"Role","type":"string"}}},"UserInfoModel":{"title":"UserInfoModel","required":["id","email","name","role","places"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"email":{"title":"Email","type":"string"},"name":{"title":"Name","type":"string"},"role":{"title":"Role","type":"string"},"places":{"title":"Places","type":"array","items":{}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}